Update per PR review suggestions
