--- learning/katas/go/Core Transforms/Map/FlatMapElements/go.mod
@@ -1,27 +1,0 @@
-// Licensed to the Apache Software Foundation (ASF) under one or more
-// contributor license agreements.  See the NOTICE file distributed with
-// this work for additional information regarding copyright ownership.
-// The ASF licenses this file to You under the Apache License, Version 2.0
-// (the "License"); you may not use this file except in compliance with
-// the License.  You may obtain a copy of the License at
-//
-//    http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-
-module flatmapelements
-
-go 1.13
-
-require (
-	github.com/apache/beam v2.20.0+incompatible
-	github.com/golang/protobuf v1.4.0 // indirect
-	github.com/googleapis/gax-go v2.0.2+incompatible // indirect
-	go.opencensus.io v0.22.3 // indirect
-	google.golang.org/api v0.22.0 // indirect
-	google.golang.org/grpc v1.29.1 // indirect
-)
