--- learning/katas/go/Core Transforms/Map/MapElements/task-info.yaml
@@ -1,38 +1,0 @@
-#
-# Licensed to the Apache Software Foundation (ASF) under one
-# or more contributor license agreements.  See the NOTICE file
-# distributed with this work for additional information
-# regarding copyright ownership.  The ASF licenses this file
-# to you under the Apache License, Version 2.0 (the
-# "License"); you may not use this file except in compliance
-# with the License.  You may obtain a copy of the License at
-#
-#  http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing,
-# software distributed under the License is distributed on an
-# "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
-# KIND, either express or implied.  See the License for the
-# specific language governing permissions and limitations
-# under the License.
-#
-
-type: edu
-files:
-- name: test/task_test.go
-  visible: false
-- name: pkg/task/task.go
-  visible: true
-  placeholders:
-  - offset: 950
-    length: 30
-    placeholder_text: TODO()
-  - offset: 1024
-    length: 123
-    placeholder_text: TODO()
-- name: cmd/main.go
-  visible: true
-- name: go.mod
-  visible: false
-- name: go.sum
-  visible: false
