+++ learning/katas/go/Core Transforms/Map/ParDo struct/go.mod
