+++ learning/katas/go/Core Transforms/Map/ParDo struct/test/task_test.go
@@ -13,21 +13,32 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package task
+package test
 
-import "github.com/apache/beam/sdks/go/pkg/beam"
+import (
+	"github.com/apache/beam/sdks/go/pkg/beam"
+	"github.com/apache/beam/sdks/go/pkg/beam/testing/passert"
+	"github.com/apache/beam/sdks/go/pkg/beam/testing/ptest"
+	"mapelements/pkg/task"
+	"testing"
+)
 
-func ApplyTransform(s beam.Scope, input beam.PCollection) beam.PCollection {
-	processFn := &multiplyByFn{
-		Factor: 5,
+func TestTask(t *testing.T) {
+	p, s := beam.NewPipelineWithRoot()
+	tests := []struct {
+		input beam.PCollection
+		want []interface{}
+	}{
+		{
+			input: beam.Create(s, 1, 2, 3, 4, 5),
+			want: []interface{}{5, 10, 15, 20, 25},
+		},
 	}
-	return beam.ParDo(s, processFn, input)
-}
-
-type multiplyByFn struct {
-	Factor int
-}
-
-func (f *multiplyByFn) ProcessElement(input int) int {
-	return f.Factor * input
+	for _, tt := range tests {
+		got := task.ApplyTransform(s, tt.input)
+		passert.Equals(s, got, tt.want...)
+		if err := ptest.Run(p); err != nil {
+			t.Error(err)
+		}
+	}
 }
